Update Jenkinsfile-trident-master
Add back in the "clean" step at the beginning of the prepare stage.
Something with the ISO generation is still not installing the proper versions of packages unless we do a full clean/rebuild of the base. Not sure what yet, but this quick fix will let us get another (albeit slower) build done while we are still looking into which cache dir needs to get flushed before starting an ISO build.